applied new rule for return and exception status message

+++ app/api/v1/endpoints/reports.py
@@ -1,7 +1,7 @@
-from fastapi import APIRouter, Depends, HTTPException, status, File, UploadFile, Form
+from fastapi import APIRouter, Depends, status, File, UploadFile, Form
 from sqlalchemy.orm import Session
 from typing import List, Optional
-from uuid import UUID, uuid4
+from uuid import UUID
 
 from app.dependencies import get_db
 from app.crud import (
@@ -21,12 +21,19 @@ from app.core.security import (
     get_current_active_admin,
     check_resource_ownership
 )
+from app.api.status_code import (
+    APIResponse,
+    SuccessMessage,
+    ErrorMessage,
+    HTTPStatus
+)
 
 from app.crud.crud_media import minio_service
 
 router = APIRouter()
 
-@router.get("/", response_model=List[Report])
+@router.get("/", response_model=APIResponse,
+            status_code=HTTPStatus.OK)
 def read_reports(
     skip: int = 0,
     limit: int = 100,
@@ -36,17 +43,37 @@ def read_reports(
     """
     Ambil semua reports
     """
-    if valid_user is not None:
-        reports = crud_report.get_reports(db=db, skip=skip, limit=limit)
-        
-        for report in reports:
-            for media in report.media:
-                media.url = minio_service.get_file_url(media.media_url, expires=7200)
-        return reports
-    else:
-        return None
+    if valid_user is None:
+        error = ErrorMessage.FORBIDDEN
+        ErrorMessage.raise_exception(
+            error_type=error,
+            details="Make sure your account have permission"
+        )
+    reports = crud_report.get_reports(db=db, skip=skip, limit=limit)
 
-@router.get("/{report_id}", response_model=Report)
+    if reports is None:
+        error = ErrorMessage.REPORT_NOT_FOUND
+        ErrorMessage.raise_exception(
+            error_type=error,
+        )
+
+    for report in reports:
+        for media in report.media:
+            media.url = minio_service.get_file_url(media.media_url, expires=7200)
+    reports_data = [Report.model_validate(report).model_dump()
+                    for report in reports]
+    
+    return APIResponse(
+        success=True,
+        message=SuccessMessage.REPORT_RETRIEVED,
+        data={
+            "reports": reports_data,
+            "total": len(reports_data)
+        }
+    )
+
+@router.get("/{report_id}", response_model=Report,
+            status_code=HTTPStatus.OK)
 def read_report(
     report_id: UUID,
     db: Session = Depends(get_db),
@@ -57,23 +84,36 @@ def read_report(
     """
     db_report = crud_report.get_report(db=db, report_id=report_id)
     if db_report is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Report not found"
+        error = ErrorMessage.REPORT_NOT_FOUND
+        ErrorMessage.raise_exception(
+            error_type=error,
         )
 
-    check_resource_ownership(
+    valid_owner = check_resource_ownership(
         resource_user_id=db_report.user_id,
         current_user=current_user,
         resource_name="report"
     )
+    
+    if valid_owner is None:
+        error = ErrorMessage.FORBIDDEN
+        ErrorMessage.raise_exception(
+            error_type=error,
+            details="Make sure your account have access permission"
+        )
 
     for media in db_report.media:
             media.url = minio_service.get_file_url(media.media_url, expires=7200)
-    return db_report
 
-@router.post("/", response_model=Report, 
-                status_code=status.HTTP_201_CREATED)
+    report_data = Report.model_validate(db_report)
+    return APIResponse(
+        success=True,
+        message=SuccessMessage.REPORT_RETRIEVED,
+        data=report_data.model_dump()
+    )
+
+@router.post("/", response_model=APIResponse,
+                status_code=HTTPStatus.CREATED)
 async def create_report(
     category: str = Form(...),
     description: Optional[str] = Form(None),
@@ -88,6 +128,12 @@ async def create_report(
     """
     Buat report baru
     """
+    if current_user is None:
+        error = ErrorMessage.UNAUTHORIZED
+        ErrorMessage.raise_exception(
+            error_type=error,
+            details="Please make sure you are logged in"
+        )
     # Ambil user id
     user_id = current_user.id
 
@@ -99,15 +145,23 @@ async def create_report(
     )
     db_location = crud_location.create_location(db=db, location=location_data)
 
+    if db_location is None:
+        error = ErrorMessage.MISSING_REQUIRED_FIELD
+        ErrorMessage.raise_exception(
+            error_type=error,
+            details="Please make sure those \"location\" fields are completed"
+        )
+
     # Cek atau ambil report status id
     status_id = None
     if report_status_id:
         try:
             status_id = UUID(report_status_id)
         except ValueError:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Invalid report_status_id format"
+            error = ErrorMessage.MISSING_REQUIRED_FIELD
+            ErrorMessage.raise_exception(
+                error_type=error,
+                details="Please make sure those \"status id\" fields are completed"
             )
     else:
         status_id = crud_report_status.get_status_id_by_name(db=db)
@@ -119,7 +173,7 @@ async def create_report(
         location=location_data,
         report_status_id=status_id
     )
-    
+
     db_report = crud_report.create_report(
         db=db,
         report=report_data,
@@ -127,31 +181,48 @@ async def create_report(
         location_id=db_location.id,
         report_status_id=status_id
     )
-    
+
+    if db_report is None:
+        error = ErrorMessage.REPORT_CREATION_FAILED
+        ErrorMessage.raise_exception(
+            error_type=error,
+        )
+
     if files:
         for file in files:
             # Skip jika file kosong
             if file.filename == '':
                 continue
-                
             # Upload ke MinIO
             object_name, media_type = minio_service.upload_file(
                 file=file,
                 folder=f"report_media/{db_report.id}"
             )
-            
-            # Simpan metadata ke database
-            crud_media.create_media(
-                db=db,
-                report_id=db_report.id,
-                media_url=object_name,
-                media_type=media_type
-            )
-    
-    db.refresh(db_report)
-    return db_report
 
-@router.patch("/{report_id}", response_model=Report)
+            # Simpan metadata ke database
+            try:
+                db_media = crud_media.create_media(
+                    db=db,
+                    report_id=db_report.id,
+                    media_url=object_name,
+                    media_type=media_type
+                )
+            except ValueError as e:
+                error = ErrorMessage.MEDIA_UPLOAD_FAILED
+                ErrorMessage.raise_exception(
+                    error_type=error,
+                )
+
+    db.refresh(db_report)
+    report_data = Report.model_validate(db_report)
+    return APIResponse(
+        success=True,
+        message=SuccessMessage.REPORT_CREATED,
+        data=report_data.model_dump()
+    )
+
+@router.patch("/{report_id}", response_model=APIResponse,
+              status_code=HTTPStatus.OK)
 def update_report(
     report_id: UUID,
     report: ReportUpdate,
@@ -161,13 +232,18 @@ def update_report(
     """
     Update report
     """
+    if current_user is None:
+        error = ErrorMessage.FORBIDDEN
+        ErrorMessage.raise_exception(
+            error_type=error,
+        )
     new_location_id = None
-    
+
     db_report = crud_report.get_report(db=db, report_id=report_id)
     if db_report is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Report not found"
+        error = ErrorMessage.REPORT_NOT_FOUND
+        ErrorMessage.raise_exception(
+            error_type=error,
         )
 
     check_resource_ownership(
@@ -182,8 +258,14 @@ def update_report(
             db=db,
             location=report.location
         )
+        if db_location is None:
+            error = ErrorMessage.INVALID_FORMAT
+            ErrorMessage.raise_exception(
+                error_type=error,
+                details="Please make sure those \"create location\" fields are completed"
+            )
         new_location_id = db_location.id
-    
+
     # Update report
     db_report = crud_report.update_report(
         db=db, 
@@ -191,13 +273,20 @@ def update_report(
         report=report,
         new_location_id=new_location_id
     )
-    
+
     if db_report is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Report not found"
+        error = ErrorMessage.VALIDATION_ERROR
+        ErrorMessage.raise_exception(
+            error_type=error,
+            details="Please make sure those \"report\" fields are completed"
         )
-    return db_report
+    
+    report_data = Report.model_validate(db_report)
+    return APIResponse(
+        success=True,
+        message=SuccessMessage.REPORT_UPDATED,
+        data=report_data
+    )
 
 @router.delete("/{report_id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_report(
@@ -208,14 +297,18 @@ def delete_report(
     """
     Hapus report
     """
+    if current_user is None:
+        error = ErrorMessage.FORBIDDEN
+        ErrorMessage.raise_exception(
+            error_type=error,
+        )
     db_report = crud_report.get_report(db=db, report_id=report_id)
     if db_report is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Report not found"
+        error = ErrorMessage.REPORT_NOT_FOUND
+        ErrorMessage.raise_exception(
+            error_type=error,
         )
-    
-    # ✅ Validasi ownership
+
     check_resource_ownership(
         resource_user_id=db_report.user_id,
         current_user=current_user,
@@ -223,11 +316,18 @@ def delete_report(
     )
     
     for media in db_report.media:
-        minio_service.delete_file(media.url)
+        if media:
+            minio_service.delete_file(media.url)
 
     success = crud_report.delete_report(db=db, report_id=report_id)
     if not success:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Report not found"
+        error = ErrorMessage.REPORT_NOT_FOUND
+        ErrorMessage.raise_exception(
+            error_type=error,
+            details="Failed to delete at \"report\""
+        )
+    else:
+        APIResponse(
+            success=True,
+            message=SuccessMessage.REPORT_DELETED,
         )
